Refactor: move `aspire_an` function to repondeur
